controller/requirements.txt
pybullet==3.1.9
numpy==1.21.2
roboticstoolbox-python==0.11.0
sympy==1.8
cvxopt==1.2.7
git+https://github.com/xArm-Developer/xArm-Python-SDK.git@release-1.8.0
git+https://github.com/DVIC-ALFRED/xarm-hand-control.git@v1.1
